Update account routes, rename Account Management page
