+++ src/pages/account/login.js
@@ -1,49 +1,48 @@
 import React from 'react';
 
 import Link from "next/link"
-import Layout from "../components/Layout"
-import AccountFormHeader from "../components/account/AccountFormHeader"
-import AccountForm from "../components/account/AccountForm"
+import Layout from "../../components/Layout"
+import AccountFormHeader from "../../components/account/AccountFormHeader"
+import AccountForm from "../../components/account/AccountForm"
 
-const Register = () => (
+const Login = () => (
 	<Layout>
 		<div className="responsive-width mx-auto mt-3">
-			<AccountFormHeader title="Register" subtitle="Sign up to edit blog posts, the menu, and more." />
+			<AccountFormHeader title="Login" subtitle="Sign in to edit blog posts, the menu, and more." />
 			<AccountForm>
 				<div className="form-group">
 					<label for="email-input">Email</label>
 					<input type="email" className="form-control" id="email-input" aria-describedby="emailHelp" placeholder="Email" />
 				</div>
 				<div className="form-group">
-					<label for="confirm-email-input">Confirm Email</label>
-					<input type="email" className="form-control" id="confirm-email-input" placeholder="Confirm Email" />
-				</div>
-				<div className="form-group">
 					<label for="password-input">Password</label>
 					<input type="password" className="form-control" id="password-input" placeholder="Password" />
 				</div>
 				<div className="form-group">
-					<label for="confirm-password-input">Confirm Password</label>
-					<input type="password" className="form-control" id="confirm-password-input" placeholder="Confirm Password" />
-				</div>
-				<div className="form-group">
-					<button type="submit" className="btn btn-green w-100 mt-2 mb-1">Register</button>
+					<button type="submit" className="btn btn-green w-100">Sign in</button>
 				</div>
 			</AccountForm>
 		</div>
 		<span className="responsive-width link-container mx-auto mt-2 text-center dark-green-text">
-			<Link href="/login">
-				<a>Already have an account? Sign in here.</a>
+			<Link href="#">
+				<a>Forgot your password?</a>
+			</Link>
+			<Link href="/account/register">
+				<a>Register as a new user</a>
 			</Link>
 		</span>
 		<style jsx>{`
 			.responsive-width {
 				width: 40%;
 			}
+			
+			.custom-border {
+				border-radius: 10px;
+			}
 
 			.link-container {
 				display: flex;
-				justify-content: center;
+				justify-content: space-between;
 			}
 
 			@media only screen and (max-width: 600px) {
@@ -84,4 +83,4 @@ const Register = () => (
 	</Layout>
 )
 
-export default Register
+export default Login
