+++ src/pages/account/register.js
@@ -1,48 +1,49 @@
 import React from 'react';
 
 import Link from "next/link"
-import Layout from "../components/Layout"
-import AccountFormHeader from "../components/account/AccountFormHeader"
-import AccountForm from "../components/account/AccountForm"
+import Layout from "../../components/Layout"
+import AccountFormHeader from "../../components/account/AccountFormHeader"
+import AccountForm from "../../components/account/AccountForm"
 
-const Login = () => (
+const Register = () => (
 	<Layout>
 		<div className="responsive-width mx-auto mt-3">
-			<AccountFormHeader title="Login" subtitle="Sign in to edit blog posts, the menu, and more." />
+			<AccountFormHeader title="Register" subtitle="Sign up to edit blog posts, the menu, and more." />
 			<AccountForm>
 				<div className="form-group">
 					<label for="email-input">Email</label>
 					<input type="email" className="form-control" id="email-input" aria-describedby="emailHelp" placeholder="Email" />
 				</div>
 				<div className="form-group">
+					<label for="confirm-email-input">Confirm Email</label>
+					<input type="email" className="form-control" id="confirm-email-input" placeholder="Confirm Email" />
+				</div>
+				<div className="form-group">
 					<label for="password-input">Password</label>
 					<input type="password" className="form-control" id="password-input" placeholder="Password" />
 				</div>
 				<div className="form-group">
-					<button type="submit" className="btn btn-green w-100">Sign in</button>
+					<label for="confirm-password-input">Confirm Password</label>
+					<input type="password" className="form-control" id="confirm-password-input" placeholder="Confirm Password" />
+				</div>
+				<div className="form-group">
+					<button type="submit" className="btn btn-green w-100 mt-2 mb-1">Register</button>
 				</div>
 			</AccountForm>
 		</div>
 		<span className="responsive-width link-container mx-auto mt-2 text-center dark-green-text">
-			<Link href="#">
-				<a>Forgot your password?</a>
-			</Link>
-			<Link href="/register">
-				<a>Register as a new user</a>
+			<Link href="/account/login">
+				<a>Already have an account? Sign in here.</a>
 			</Link>
 		</span>
 		<style jsx>{`
 			.responsive-width {
 				width: 40%;
 			}
-			
-			.custom-border {
-				border-radius: 10px;
-			}
 
 			.link-container {
 				display: flex;
-				justify-content: space-between;
+				justify-content: center;
 			}
 
 			@media only screen and (max-width: 600px) {
@@ -83,4 +84,4 @@ const Login = () => (
 	</Layout>
 )
 
-export default Login
+export default Register
